Adds GTraversable instance for Generic types
The instance is for types where every field is significant (and
therefore has the constraint hold). In this case, it's possible to write
a traversal over all the fields using the conversion to/from the Generic
representation.
The only questionable part of this design is the GTraversable' instance
for composition of functors.